pushing the latest code

+++ firmware/src/main/java/fm/repository/PropertyRepository.java
@@ -3,5 +3,5 @@ package fm.repository;
 import fm.entity.PropertyEntity;
 import org.springframework.data.r2dbc.repository.R2dbcRepository;
 
-public interface PropertyRepository extends R2dbcRepository<PropertyEntity,Integer> {
+public interface PropertyRepository extends R2dbcRepository<PropertyEntity,Long> {
 }
